Add support for drawing various geometries
The leaflet engine now supports drawing the various geometry types from
Montage-Geo.  This includes normalizing the geometry's coordinates as
the map moves from world copy to world copy.  Geometries that cross the
anti-meridian are also normalized to compenstate for a limitation with
Leaflet.

A sample component was created that demonstrates drawing the countries
of the world.  The countries are imported via local geoJSON file.