Adds cluster organizer, starts feature overlay
The cluster organizer sorts a set of features into clusters.  This imp-
lementation is a variation of the marker clustering found in the Leaflet
open source project.  This implementation is proven to be fast at any
zoom level for a large number of features.  Importantly, increasing the
zoom level does not impose a performance penalty.

Starts the feature overlay component.

Adds a number of methods to the point object.

Improves documentation in many components and objects.

Adds a number of tests.

Moves the sample components into an examples directory inside of the ui
folder.