Add Heatmap Overlay and embedded overlays
First stab at allowing other components to be embedded inside the map.
The map has a proprety called overlays which is an array.  The owner
passes the array of overlays that are meant to be added to the map.  The
map in turn will ensure that the overlay is embedded into the engine.
Once the overlay is embedded the map will pass itself to the overlay.
At that point the overlay can initialize itself.